refactor: migrate routing to Next.js App Router, remove react-router usage
