+++ app/globals.css
@@ -28,6 +28,7 @@ body {
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
+  justify-content: center;
 }
 
 h1 {
@@ -74,4 +75,325 @@ button:focus-visible {
 [data-theme='light'] {
   background: #fff;
   color: #222;
+}
+
+/* главная */
+
+.selected-flyout {
+  position: fixed;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: #222;
+  color: #fff;
+  padding: 16px;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+  text-align: center;
+}
+
+.selected-flyout span {
+  margin-right: 16px;
+}
+
+.selected-flyout button {
+  margin-top: 0;
+  margin-right: 8px;
+}
+
+.selected-flyout button:last-child {
+  margin-right: 0;
+}
+
+body {
+  margin: 0;
+  font-family: Arial, sans-serif;
+}
+
+#root {
+  max-width: 1280px;
+  margin: 0 auto 80px auto;
+  padding: 2rem;
+  text-align: center;
+}
+
+header {
+  background-color: #282c34;
+  padding: 20px;
+  color: white;
+  text-align: center;
+  position: sticky;
+  top: 0;
+  z-index: 1000;
+}
+
+header h1 {
+  margin: 0;
+}
+
+header .search-bar {
+  margin-top: 20px;
+}
+
+main {
+  padding: 20px;
+}
+
+.card-list {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px;
+  justify-content: center;
+}
+
+.card {
+  border: 1px solid #ccc;
+  padding: 20px;
+  border-radius: 8px;
+  width: 200px;
+  text-align: center;
+}
+
+button {
+  margin-top: 20px;
+}
+
+.logo {
+  height: 6em;
+  padding: 1.5em;
+  will-change: filter;
+  transition: filter 300ms;
+}
+
+.logo:hover {
+  filter: drop-shadow(0 0 2em #646cffaa);
+}
+
+.logo.react:hover {
+  filter: drop-shadow(0 0 2em #61dafbaa);
+}
+
+.read-the-docs {
+  color: #888;
+}
+
+
+/* ---About page-- */
+.about-container {
+  max-width: 800px;
+  margin: 40px auto;
+  padding: 30px;
+  background-color: #ffffff;
+  border-radius: 16px;
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.08);
+  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+  color: #333;
+  text-align: left;
+}
+
+.about-container section + section {
+  margin-top: 32px;
+}
+
+.about-container h2 {
+  font-size: 28px;
+  color: #222;
+  margin-bottom: 12px;
+}
+
+.about-container h3 {
+  font-size: 22px;
+  color: #444;
+  margin-bottom: 8px;
+}
+
+.about-container p {
+  font-size: 16px;
+  line-height: 1.6;
+}
+
+.about-container a {
+  color: #0077cc;
+  text-decoration: none;
+}
+
+.about-container a:hover {
+  text-decoration: underline;
+}
+
+.about-container ul {
+  list-style: none;
+  padding-left: 0;
+  font-size: 16px;
+}
+
+.about-container li {
+  margin-bottom: 6px;
+}
+
+.skills-list {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(180px, 1fr));
+  gap: 6px 20px;
+}
+
+.about-container pre {
+  background-color: #f4f4f4;
+  padding: 16px;
+  border-radius: 8px;
+  overflow-x: auto;
+  font-family: 'Courier New', Courier, monospace;
+  font-size: 14px;
+  line-height: 1.5;
+}
+
+[data-theme='dark'] .about-container {
+  background-color: #232323;
+  color: #eee;
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.32);
+}
+
+[data-theme='dark'] .about-container h2,
+[data-theme='dark'] .about-container h3 {
+  color: #fff;
+}
+
+[data-theme='dark'] .about-container pre {
+  background-color: #181818;
+  color: #ccc;
+}
+
+[data-theme='dark'] .about-container a {
+  color: #6ab7ff;
+}
+
+.card-list {
+  margin: 50px 0;
+}
+
+.navigation-top {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  padding: 10px 0;
+}
+
+.navigation-top button {
+margin: 0;
+}
+
+.pagination {
+  display: flex;
+  justify-content: center;
+  margin: 20px 20px 70px;
+}
+
+/* searchBar */
+.search-button {
+  background-color: #3498db;
+  color: white;
+  border: none;
+  padding: 10px 20px;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+  margin-left: 10px;
+}
+
+.search-button:hover {
+  background-color: #2980b9;
+}
+
+/* result */
+
+.throw-error-button {
+  background-color: #3498db;
+  color: white;
+  border: none;
+  padding: 10px 20px;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+}
+
+.throw-error-button:hover {
+  background-color: #2980b9;
+}
+
+.results-container {
+  margin-top: 32px;
+  margin-bottom: 32px;
+}
+
+/* ErroreBoundary  */
+.go-back-button {
+  background-color: #3498db;
+  color: white;
+  border: none;
+  padding: 10px 20px;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+}
+
+.go-back-button:hover {
+  background-color: #2980b9;
+}
+
+/* spinner */
+
+.spinner-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: fixed;
+  top: 54%;
+  left: 0;
+  width: 100%;
+  z-index: 9999;
+}
+
+.spinner {
+  width: 40px;
+  height: 40px;
+  border: 4px solid #f3f3f3;
+  border-top: 4px solid #3498db;
+  border-radius: 50%;
+  animation: spin 2s linear infinite;
+}
+
+.double-bounce1,
+.double-bounce2 {
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+  background-color: #333;
+  opacity: 0.6;
+  position: absolute;
+  animation: sk-bounce 2s infinite ease-in-out;
+}
+
+.double-bounce2 {
+  animation-delay: -1s;
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+@keyframes sk-bounce {
+  0%,
+  100% {
+    transform: scale(0);
+  }
+  50% {
+    transform: scale(1);
+  }
 }
