+++ app/layout.tsx
@@ -1,10 +1,17 @@
-import './globals.css';
-import { ReactNode } from 'react';
+// import '../globals.css';
+import ReduxProvider from '../src/store/ReduxProvider';
+import ThemeProvider from '../src/context/ThemeContext';
 
-export default function RootLayout({ children }: { children: ReactNode }) {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body>
+        <ReduxProvider>
+          <ThemeProvider>
+            {children}
+          </ThemeProvider>
+        </ReduxProvider>
+      </body>
     </html>
   );
 }
