--- src/App.css
@@ -1,71 +1,0 @@
-body {
-  margin: 0;
-  font-family: Arial, sans-serif;
-}
-
-#root {
-  max-width: 1280px;
-  margin: 0 auto 80px auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-header {
-  background-color: #282c34;
-  padding: 20px;
-  color: white;
-  text-align: center;
-  position: sticky;
-  top: 0;
-  z-index: 1000;
-}
-
-header h1 {
-  margin: 0;
-}
-
-header .search-bar {
-  margin-top: 20px;
-}
-
-main {
-  padding: 20px;
-}
-
-.card-list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-  justify-content: center;
-}
-
-.card {
-  border: 1px solid #ccc;
-  padding: 20px;
-  border-radius: 8px;
-  width: 200px;
-  text-align: center;
-}
-
-button {
-  margin-top: 20px;
-}
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-.read-the-docs {
-  color: #888;
-}
