--- src/components/CardList/CardList.css
@@ -1,3 +1,0 @@
-.card-list {
-  margin: 50px 0;
-}
