--- src/components/ErrorBoundary/ErrorBoundary.css
@@ -1,13 +1,0 @@
-.go-back-button {
-  background-color: #3498db;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.go-back-button:hover {
-  background-color: #2980b9;
-}
