--- src/components/Results/Results.css
@@ -1,18 +1,0 @@
-.throw-error-button {
-  background-color: #3498db;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.throw-error-button:hover {
-  background-color: #2980b9;
-}
-
-.results-container {
-  margin-top: 32px;
-  margin-bottom: 32px;
-}
