+++ src/components/Results/Results.tsx
@@ -1,4 +1,3 @@
-import './Results.css';
 import CardList from '../CardList/CardList';
 
 interface ResultsProps {
