--- src/components/SearchBar/SearchBar.css
@@ -1,14 +1,0 @@
-.search-button {
-  background-color: #3498db;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  margin-left: 10px;
-}
-
-.search-button:hover {
-  background-color: #2980b9;
-}
