--- src/components/SelectedFlyout/SelectedFlyout.css
@@ -1,25 +1,0 @@
-.selected-flyout {
-  position: fixed;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: #222;
-  color: #fff;
-  padding: 16px;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-.selected-flyout span {
-  margin-right: 16px;
-}
-
-.selected-flyout button {
-  margin-top: 0;
-  margin-right: 8px;
-}
-
-.selected-flyout button:last-child {
-  margin-right: 0;
-}
