+++ src/components/Spinner/Spinner.tsx
@@ -1,5 +1,3 @@
-import './Spinner.css';
-
 const Spinner = () => (
   <div className="spinner-container">
     <div className="spinner">
