+++ src/context/useTheme.ts
@@ -1,4 +1,6 @@
 import { useContext } from 'react';
 import { ThemeContext } from './ThemeContextContext';
 
-export const useTheme = () => useContext(ThemeContext);
+export function useTheme() {
+  return useContext(ThemeContext);
+}
