--- src/main.tsx
@@ -1,28 +1,0 @@
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
-import { Provider } from 'react-redux';
-import { store } from './store';
-import './index.css';
-import App from './App';
-import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
-import { ThemeProvider } from './context/ThemeContext';
-
-const rootElement = document.getElementById('root');
-if (rootElement) {
-  createRoot(rootElement).render(
-    <StrictMode>
-      <ErrorBoundary>
-        <ThemeProvider>
-          <Provider store={store}>
-            <BrowserRouter>
-              <App />
-            </BrowserRouter>
-          </Provider>
-        </ThemeProvider>
-      </ErrorBoundary>
-    </StrictMode>
-  );
-} else {
-  console.error('Root element not found');
-}
