+++ src/store/ReduxProvider.tsx
@@ -1,0 +1,7 @@
+'use client';
+import { Provider } from 'react-redux';
+import { store } from './store';
+
+export default function ReduxProvider({ children }: { children: React.ReactNode }) {
+  return <Provider store={store}>{children}</Provider>;
+}
