+++ tsconfig.json
@@ -35,6 +35,7 @@
   },
   "include": [
     "next-env.d.ts",
+    "global.d.ts", 
     ".next/types/**/*.ts",
     "**/*.ts",
     "**/*.tsx"
